Add Hangfire to every app
Also moves the UI to the SupportUi and makes the Worker the Hangfire
server.
