+++ TeachingRecordSystem/Directory.Packages.props
@@ -21,6 +21,7 @@
     <PackageVersion Include="GovUk.Frontend.AspNetCore" Version="1.4.0" />
     <PackageVersion Include="Hangfire.AspNetCore" Version="1.8.6" />
     <PackageVersion Include="Hangfire.Core" Version="1.8.6" />
+    <PackageVersion Include="Hangfire.NetCore" Version="1.8.6" />
     <PackageVersion Include="Hangfire.PostgreSql" Version="1.20.4" />
     <PackageVersion Include="Humanizer.Core" Version="2.14.1" />
     <PackageVersion Include="IdentityModel" Version="6.2.0" />
