+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/Infrastructure/Redis/ServiceCollectionExtensions.cs
@@ -6,8 +6,7 @@ public static class ServiceCollectionExtensions
     public static IServiceCollection AddRedis(
         this IServiceCollection services,
         IWebHostEnvironment environment,
-        IConfiguration configuration,
-        IHealthChecksBuilder healthChecksBuilder)
+        IConfiguration configuration)
     {
         if (environment.IsProduction())
         {
@@ -16,7 +15,7 @@ public static class ServiceCollectionExtensions
             services.AddSingleton<IConnectionMultiplexer>(ConnectionMultiplexer.Connect(connectionString));
             services.AddStackExchangeRedisCache(options => options.Configuration = connectionString);
 
-            healthChecksBuilder.AddRedis(connectionString);
+            services.AddHealthChecks().AddRedis(connectionString);
         }
 
         return services;
