+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/Infrastructure/Security/AuthorizationPolicies.cs
@@ -4,7 +4,6 @@ public static class AuthorizationPolicies
 {
     public const string ApiKey = nameof(ApiKey);
     public const string IdentityUserWithTrn = nameof(IdentityUserWithTrn);
-    public const string Hangfire = nameof(Hangfire);
     public const string GetPerson = nameof(GetPerson);
     public const string UpdatePerson = nameof(UpdatePerson);
     public const string UpdateNpq = nameof(UpdateNpq);
@@ -15,16 +14,13 @@ public static class AuthorizationPolicies
         switch (policy)
         {
             case ApiKey:
-            case Hangfire:
             case GetPerson:
             case UpdatePerson:
             case UpdateNpq:
             case UnlockPerson:
                 return true;
-            case IdentityUserWithTrn:
             default:
                 return false;
-
         }
     }
 }
