+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/Program.cs
@@ -1,8 +1,6 @@
 using System.Security.Claims;
 using FluentValidation;
 using FluentValidation.AspNetCore;
-using Hangfire;
-using Hangfire.PostgreSql;
 using idunno.Authentication.Basic;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Mvc.ApiExplorer;
@@ -19,7 +17,6 @@ using TeachingRecordSystem.Api.Infrastructure.RateLimiting;
 using TeachingRecordSystem.Api.Infrastructure.Redis;
 using TeachingRecordSystem.Api.Infrastructure.Security;
 using TeachingRecordSystem.Api.Validation;
-using TeachingRecordSystem.Core.DataStore.Postgres;
 using TeachingRecordSystem.Core.Dqt;
 using TeachingRecordSystem.Core.Infrastructure;
 using TeachingRecordSystem.Core.Jobs;
@@ -47,8 +44,6 @@ public class Program
         var configuration = builder.Configuration;
 
         builder.ConfigureLogging();
-
-        string pgConnectionString = configuration.GetRequiredValue("ConnectionStrings:DefaultConnection");
 
         services.AddAuthentication(ApiKeyAuthenticationHandler.AuthenticationScheme)
             .AddScheme<ApiKeyAuthenticationOptions, ApiKeyAuthenticationHandler>(ApiKeyAuthenticationHandler.AuthenticationScheme, _ => { })
@@ -103,22 +98,15 @@ public class Program
                     .RequireClaim(ClaimTypes.Name));
 
             options.AddPolicy(
-                    AuthorizationPolicies.IdentityUserWithTrn,
-                    policy => policy
-                        .AddAuthenticationSchemes(JwtBearerDefaults.AuthenticationScheme)
-                        .RequireAssertion(ctx =>
-                        {
-                            var scopes = (ctx.User.FindFirstValue("scope") ?? string.Empty).Split(' ', StringSplitOptions.RemoveEmptyEntries);
-                            return scopes.Contains("dqt:read");
-                        })
-                        .RequireClaim("trn"));
-
-            options.AddPolicy(
-                AuthorizationPolicies.Hangfire,
+                AuthorizationPolicies.IdentityUserWithTrn,
                 policy => policy
-                    .AddAuthenticationSchemes(BasicAuthenticationDefaults.AuthenticationScheme)
-                    .RequireAuthenticatedUser()
-                );
+                    .AddAuthenticationSchemes(JwtBearerDefaults.AuthenticationScheme)
+                    .RequireAssertion(ctx =>
+                    {
+                        var scopes = (ctx.User.FindFirstValue("scope") ?? string.Empty).Split(' ', StringSplitOptions.RemoveEmptyEntries);
+                        return scopes.Contains("dqt:read");
+                    })
+                    .RequireClaim("trn"));
 
             options.AddPolicy(
                 AuthorizationPolicies.GetPerson,
@@ -182,9 +170,6 @@ public class Program
 
         services.AddOpenApi(configuration);
 
-        var healthCheckBuilder = services.AddHealthChecks()
-            .AddNpgSql(pgConnectionString);
-
         services.AddMediatR(cfg => cfg.RegisterServicesFromAssemblyContaining<Program>());
         services.AddSingleton<IApiClientRepository, ConfigurationApiClientRepository>();
         services.AddSingleton<ICurrentClientProvider, ClaimsPrincipalCurrentClientProvider>();
@@ -197,29 +182,9 @@ public class Program
             client.Timeout = TimeSpan.FromSeconds(30);
         });
 
-        services.AddDbContext<TrsDbContext>(
-            options => TrsDbContext.ConfigureOptions(options, pgConnectionString),
-            contextLifetime: ServiceLifetime.Transient,
-            optionsLifetime: ServiceLifetime.Singleton);
-
-        services.AddDbContextFactory<TrsDbContext>(options => TrsDbContext.ConfigureOptions(options, pgConnectionString));
-
-        services.AddDatabaseDeveloperPageExceptionFilter();
-
         builder.AddBlobStorage();
 
         builder.AddDistributedLocks();
-
-        if (!env.IsUnitTests() && !env.IsEndToEndTests())
-        {
-            services.AddHangfire(configuration => configuration
-                .SetDataCompatibilityLevel(CompatibilityLevel.Version_170)
-                .UseSimpleAssemblyNameTypeSerializer()
-                .UseRecommendedSerializerSettings()
-                .UsePostgreSqlStorage(o => o.UseNpgsqlConnection(pgConnectionString)));
-
-            services.AddHangfireServer();
-        }
 
         services.AddTrnGenerationApi(configuration);
         services.AddIdentityApi(configuration, env);
@@ -236,10 +201,11 @@ public class Program
             services.AddDefaultServiceClient(ServiceLifetime.Transient, _ => crmServiceClient.Clone());
             services.AddTransient<IDataverseAdapter, DataverseAdapter>();
 
-            healthCheckBuilder.AddCheck("CRM", () => crmServiceClient.IsReady ? HealthCheckResult.Healthy() : HealthCheckResult.Degraded());
+            services.AddHealthChecks()
+                .AddCheck("CRM", () => crmServiceClient.IsReady ? HealthCheckResult.Healthy() : HealthCheckResult.Degraded());
         }
 
-        services.AddRedis(env, configuration, healthCheckBuilder);
+        services.AddRedis(env, configuration);
         services.AddRateLimiting(env, configuration);
 
         var app = builder.Build();
@@ -270,11 +236,6 @@ public class Program
         app.MapWebHookEndpoints();
 
         app.MapControllers();
-
-        if (!builder.Environment.IsUnitTests() && !builder.Environment.IsEndToEndTests())
-        {
-            app.MapHangfireDashboardWithAuthorizationPolicy(AuthorizationPolicies.Hangfire, "/_hangfire");
-        }
 
         if (env.IsDevelopment())
         {
