+++ TeachingRecordSystem/src/TeachingRecordSystem.Hosting/Extensions.cs
@@ -1,0 +1,72 @@
+using Hangfire;
+using Hangfire.PostgreSql;
+using Microsoft.ApplicationInsights.Extensibility;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Npgsql;
+using Serilog;
+using Serilog.Formatting.Compact;
+using TeachingRecordSystem.Core;
+using TeachingRecordSystem.Core.DataStore.Postgres;
+
+namespace TeachingRecordSystem.Hosting;
+
+public static class Extensions
+{
+    public static IHostApplicationBuilder AddDatabase(this IHostApplicationBuilder builder)
+    {
+        var pgConnectionString = GetPostgresConnectionString(builder.Configuration);
+
+        builder.Services.AddDbContext<TrsDbContext>(
+            options => TrsDbContext.ConfigureOptions(options, pgConnectionString),
+            contextLifetime: ServiceLifetime.Transient,
+            optionsLifetime: ServiceLifetime.Singleton);
+
+        builder.Services.AddDbContextFactory<TrsDbContext>(options => TrsDbContext.ConfigureOptions(options, pgConnectionString));
+
+        return builder;
+    }
+
+    public static void AddHangfire(this IHostApplicationBuilder builder)
+    {
+        if (!builder.Environment.IsUnitTests() && !builder.Environment.IsEndToEndTests())
+        {
+            var pgConnectionString = GetPostgresConnectionString(builder.Configuration);
+
+            builder.Services.AddHangfire(configuration => configuration
+                .SetDataCompatibilityLevel(CompatibilityLevel.Version_170)
+                .UseSimpleAssemblyNameTypeSerializer()
+                .UseRecommendedSerializerSettings()
+                .UsePostgreSqlStorage(o => o.UseNpgsqlConnection(pgConnectionString)));
+        }
+    }
+
+    public static void ConfigureSerilog(
+        this LoggerConfiguration config,
+        IHostEnvironment environment,
+        IConfiguration configuration,
+        IServiceProvider services)
+    {
+        config
+            .ReadFrom.Configuration(configuration)
+            .WriteTo.ApplicationInsights(services.GetRequiredService<TelemetryConfiguration>(), TelemetryConverter.Traces)
+            .WriteTo.Sentry(o => o.InitializeSdk = false);
+
+        if (environment.IsProduction())
+        {
+            config.WriteTo.Console(new CompactJsonFormatter());
+        }
+        else
+        {
+            config.WriteTo.Console();
+        }
+    }
+
+    public static string GetPostgresConnectionString(this IConfiguration configuration) =>
+        new NpgsqlConnectionStringBuilder(configuration.GetRequiredValue("ConnectionStrings:DefaultConnection"))
+        {
+            // We rely on error details to get the offending duplicate key values in the TrsDataSyncHelper
+            IncludeErrorDetail = true
+        }.ConnectionString;
+}
