--- TeachingRecordSystem/src/TeachingRecordSystem.Hosting/HostApplicationBuilderExtensions.cs
@@ -1,32 +1,0 @@
-using Microsoft.ApplicationInsights.Extensibility;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Serilog;
-using Serilog.Formatting.Compact;
-
-namespace TeachingRecordSystem.Hosting;
-
-public static class Extensions
-{
-    public static void ConfigureSerilog(
-        this LoggerConfiguration config,
-        IHostEnvironment environment,
-        IConfiguration configuration,
-        IServiceProvider services)
-    {
-        config
-            .ReadFrom.Configuration(configuration)
-            .WriteTo.ApplicationInsights(services.GetRequiredService<TelemetryConfiguration>(), TelemetryConverter.Traces)
-            .WriteTo.Sentry(o => o.InitializeSdk = false);
-
-        if (environment.IsProduction())
-        {
-            config.WriteTo.Console(new CompactJsonFormatter());
-        }
-        else
-        {
-            config.WriteTo.Console();
-        }
-    }
-}
