+++ TeachingRecordSystem/src/TeachingRecordSystem.SupportUi/Infrastructure/Security/AuthorizationPolicies.cs
@@ -3,4 +3,5 @@ namespace TeachingRecordSystem.SupportUi.Infrastructure.Security;
 public static class AuthorizationPolicies
 {
     public const string CaseManagement = "CaseManagement";
+    public const string Hangfire = "Hangfire";
 }
